requirements.txt
# Core LlamaIndex packages
llama-index-core>=0.12.0
llama-index-embeddings-openai>=0.3.0
llama-index-llms-openai>=0.3.0
llama-index-vector-stores-chroma>=0.3.0

# Vector store
chromadb>=0.5.0

# OpenAI API
openai>=1.0.0

# PDF processing
PyMuPDF>=1.24.0

# Data validation
pydantic>=2.0.0

# Environment management
python-dotenv>=1.0.0